Overhaul the conversion traits
This commit overhauls the conversion traits used for types crossing the Rust/JS
boundary. Previously there were a few ad-hoc traits but now there's only two.
One is for converting JS values to Rust values, `FromWasmAbi`, and the other is
for converting Rust values into JS values, `IntoWasmAbi`. These two traits are
now uniformly used for all operations such as creating references and such.

The `IntoWasmAbi` trait is pretty straightforward but `FromWasmAbi` is much
trickier. We need to be able to create references which necessitates a good deal
of funkiness with the API, but it in general should work out alright.

The goal of this commit is to reduce the number of traits in play so it's more
straightforward which traits are used for closures.