Add `JsValue::from_wasm_abi`
This commit implements a generic function on `JsValue` which can convert
*any* Rust value that implements `IntoWasmAbi` into the corresponding
`JsValue`. This can be handy when binding APIs and the intermediate JS
value is needed is needed over a few function calls or is otherwise
needed when calling APIs that take `JsValue` but otherwise work with
Rust primitives.

One of the primary intended use cases for this will be interoperation
with the various `JsValue`-taking functions (or `Object`) in `web-sys`.

cc #811